*: replaced godep with glide
Godep was continually being a drain on development resources of acbuild,
and thus this commit replaces godep with glide. Glide takes advantage of
the golang vendor experiment, and thus this change will drop support for
versions of Go less than 1.5.

The biggest impact this change has is that dependencies are no longer
vendored with acbuild, and are instead fetched at compile time by the
build script. This is done by checking if the `vendor` directory exists,
and running `glide install` if it is not.