*: store persists between builds, added commands to interact with store
This commit modifies acbuild to retain the images it fetches between
builds. These images are kept in `/var/lib/acbuild`. This results in a
better UX, as iterating on an acbuild script takes much less time when
the dependencies are not constantly re-downloaded.

The commands added to interact with the store are:

- __acbuild store fetch IMAGE__: fetches an image into the store
- __acbuild store rm IMAGE__: removes an image from the store
- __acbuild store list__: lists all images in the store, in the form of
  NAME:VERSION
- __acbuild store clear__: deletes all images in the store